Fix cppcheck warnings, typos, style issues (#1227)
Reducing CppCheck report size, frequent warnings:
* missing override (CI-clang seems also affected by this)
* different argument name
* variable scope can be reduced
* Ogre or containers in standard library most of times use size_t, better to use
* Fix typos
* In general tried to follow the rule from c++ core guidelines https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#c128-virtual-functions-should-specify-exactly-one-of-virtual-override-or-final (but changing it in all code could make PR much bigger)

(cherry picked from commit a395a9527c83dffcdedb376575d80bc00c0c4661)

+++ examples/lidar_visual/GlutWindow.cc
@@ -460,7 +460,7 @@ void run(std::vector<ir::CameraPtr> _cameras,
   initContext();
   printUsage();
   g_lidarData.clear();
-  for (int pt =0; pt < _pts.size(); pt++)
+  for (size_t pt =0; pt < _pts.size(); pt++)
   {
     g_lidarData.push_back(_pts[pt]);
   }
